app/core/gimpdrawable-blend.c some progress cleanup.
2004-08-11  Michael Natterer  <mitch@gimp.org>

	* app/core/gimpdrawable-blend.c
	* app/core/gimpprogress.c: some progress cleanup.

	* app/display/gimpstatusbar.c (gimp_statusbar_progress_start): no
	need to warn if there is already a progress active, just silently
	return NULL as all other GimpProgressInterface implementors.

	* app/plug-in/plug-in-progress.c: several progress fixes.
	It's still a mess.

	* plug-ins/common/url.c: don't show progress depending on
	run_mode. Run the actual file plug-in with the same run_mode we
	were invoked with.

+++ app/core/gimpprogress.c
@@ -189,7 +189,7 @@ gimp_progress_update_and_flush (gint     min,
                                 gpointer data)
 {
   gimp_progress_set_value (GIMP_PROGRESS (data),
-                           (gfloat) (current - min) / (gfloat) (max - min));
+                           (gdouble) (current - min) / (gdouble) (max - min));
 
   while (g_main_context_pending (NULL))
     g_main_context_iteration (NULL, TRUE);
